refactor: reuse shared queries for variation prices
